refactor : 컨펌 모달 훅 shared 로 변경 리턴 타입 obj to list 이름 useHandlerModal로 모달 온오프 이벤트 통합
